Rework polling colleciton to ping faster
